[HUGE] rework the way callbacks are dispatched
Previously, I had attached the event-juggling semaphores to the session object.
This was unsustainable as the session object was not reachable from everywhere,
and callbacks for other objects also needed these semaphores.

I made the event semaphores (and the struct where event data is stored) GLOBAL.

I also had the issue of where to dispatch callbacks. Once I have an event, how
do I know which object is to be called? The only way is to have each event store
which object it belongs to. This means I can have different handlers for each
playlist, for example.

Yeah, huge commit. Sorry to whoever happens to read this in the future. If you
are myself, please send me back some scrambled egg, bacon and toast using your
time machines. Thanks.