Update example show_public_playlists_of_user
I’ve noticed it seems to segfault if used enough times; like if you
list playlists from a lot of different users it crashes on

sp_session_process_events

Not sure why yet. On inspection, the pointer still seems to be valid;
it could very well be a bug in libspotify.