(big commit) One Big Callback™ support (see #61)
There’s a bunch of new things in this commit, and lots of things that
could potentially go wrong. Hopefully, most of the screwups that are
possible will occur in ruby-land and as such won’t mean segfaults but
mere exceptions raised.

Callbacks are now created once, and only once. When giving callbacks
to libspotify it will *always* be the same callbacks. Their handlers
are specified once (see lib/hallon/observable/<class>.rb), and when
triggered the callback will iterate through all objects interested in
receiving callback notifications and trigger the appropriate event on
them.

How this works is pretty fucking cool. When objects are created, we want
to attach callbacks to them. The callbacks can never be GC’d, which means
that we store them globally in the class of the newly created object. On
retrieval of these callbacks, it will *always* be the same object, which
means that, for example:

  Hallon::Image.callbacks.object_id == Hallon::Image.callbacks.object_id

This is important. Why, you ask? When we create an object, or wrap an
already existing pointer, we have no idea if we’ve attached callbacks to
that specific pointer before. We can’t attach any such information to
the pointer object itself, because it could be re-created with the same
address from somewhere else (for example #playlist_added_callback).

Now, to assure we don’t ever attach the same callback to the same object
twice (and such, triggering multiple callbacks for the same event), we
try to remove any callbacks from the object *before* we attach them. At
worst, we’ll simply detach and attach the same callback. You can look at
it as if it was an invisible counter we always decrement before we try to
increment it, but if it’s already at 0 it’ll stay at 0.

Anyhow. After we’ve attached the callbacks, and have a pointer reference
that we can use, we register the Hallon object as interested in any events
that might occur on its’ pointer. When we do this, we only store a *weak*
reference to the subscribee, so we don’t try to retain it in memory forever.

(also, at the same time we attach a finalizer to the subscribee, so when
 it is garbage collected we remove it from the list of subscribers for
 that pointer address — when the list is empty, we clean it away from
 memory, protecting us from leaking an empty array for each new pointer
 address)

When a callback fires, we use its’ subject pointer to find all objects
that have subscribed to events for that pointers’ address. We then simply
traverse this list, convert the weak references to actual objects and
trigger the desired event on it.